refactor(store): manage platform datas seperately
